Add identifier suffixes to spreadsheet output
This ensures that two keys pointing to the same schema definition are
handled correctly.

Adds a json_key paramater to JSONParser.parse_json_dict so that we know
what the key used to refer to this json_dict was.

Also adds the .name property to a Sheet object, so that we can later use
this information to omit the suffix if it is the same as the sheet name.