[#85] Main sheet use schema and allow json pointer
Adds the following

* A new method to unflatten the main page.
* Adds flattened property to Schema parser that more reflects json
 pointer removes
* Use schema to convert types

The new unflatten should produce output the same as before meaning that
multiple sheets will still work.
If no schema is supplied then fallback to old behaviour with types.

Various tests had to be fixed as if the schema says the field is a
string then it converts it.

Breaking change: If schema is supplied and the main table uses ":type"
notation then it will mistake this is a subobject not a types.